refactor: id card entity 폴더 정리
