+++ Domain/src/main/java/com/dingdong/domain/domains/idcard/domain/vo/Character.java
@@ -1,4 +1,4 @@
-package com.dingdong.domain.domains.idcard.domain;
+package com.dingdong.domain.domains.idcard.domain.vo;
 
 
 import com.dingdong.domain.domains.idcard.domain.enums.CharacterType;
@@ -6,9 +6,11 @@ import javax.persistence.Embeddable;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Embeddable
 @Getter
+@NoArgsConstructor
 public class Character {
 
     @Enumerated(EnumType.STRING)
